Revert CompCertX program holes
In CertiKOS we make sure the programs follow a predetermined structure so that
the resulting global environments have a fixed symbol -> block mapping. To make
that possible, in CompCertX we allow empty declarations in programs.

Instead, for Composable Compcert the plan is to use the linking infrastructure
to add external declarations as needed, so we can drop the corresponding
changes to Compcert. This commit may be incomplete as not every pass builds at
the moment.