Use blocks to identify entry point in language interfaces
There are aspects to this change.

First, at the assemby level we will no longer have function identifiers
in queries, but simply the current value of the program counter.
Relating this to the C language interface would require the global
environments to be involved if we have function names in C-level
queries. If instead we use the block from the beginning to identify the
function being called, there is no problem.

Second, there is a discrepancy between the way Compcert names internal
functions (using AST.ident), and external function (using string).
Before, as asssumed there was some mapping str2ident and used that to
fill the gap. Now, we identify external function by their declaration in
the program (so, the corresponding block, although this could have been
the corresponding AST.ident in the old interface as well).