Asm: restrict initial PC more, initial RA less
We want to restrict the PC to executing internal function only
(disallowing non-[EF_external] external functions, unlike the
[internal_pc] predicate). This is to avoid determinism issues when
linking several assembly modules: we want the sets of queries they
handle to be disjoint, which they aren't if they have common
"internal-external" function definitions.

On the other hand, we can't ensure RA points outside the module through
the calling convention, because [match_query] does not have access to
the global environment, so we remove that constraint in [initial_state].
This is going to make it tricky for the [final_state] part of the
simulation, because we don't want too many final states (ie. returing in
the middle of the execution because the initial RA pointed there), but
we also don't want foo few (ie. making it impossible to return because
RA points inside of the module -- this is how the definition is written
as of this commit). Thankfully SP should be helpful here.