Implement a decision procedure for frame_agree_perms
The challenge is to show the decidability of a predicate of the form
(forall o k p, perm m b o k p -> P o).

This is challenging because the o is a Z, therefore unbounded.
If we keep in the memory some information about the range of offsets
that may have permissions, this becomes decidable because all we have
to enumerate is that specific range.

This is implemented as a [mem_bounds] field of the memory implementation,
which associates with each block a range [lo,hi] of offsets where we /may/
have permissions. Those permissions may be dropped at some later point in
the program.

Prior to this commit, we needed to have a specific implementation of
[record_stack_blocks_none] which took as a parameter a proof that all
the offsets with permissions in the blocks of the frame that we try to
record indeed satisfied `P` (in the example above), i.e. it had to be in
the bounds of the frame itself.

With this commit, it will become useless to have both inductive and
operational definitions of [record_stack_blocks].