Fix the definition of horizontal composition
When one of the components interacts with the environment, the
environment may then want to resume the other module. This means that
we need to recombine the interaction's continuation with the suspended
continuation of the other component, similarly to how the entry
continuations are initially combined.

One consequence of this is that we need to assign a query domain not
just to the module as a whole, but to every continuation. The definition
of RTS behaviors is modified accordingly.

Taking all this into account, I restructure horizontal composition
into: 1. a flat composition operator that runs the two components side
by side and handles this combination of continuations to switch between
the components as needed; 2. a resolution operator which allows the
combined component to interact with itself, by making sure that queries
made to the environment are immediately diverted to the component's own
continuation whenever possible, so that the external interaction is
replaced with an internal action.